feat: use new relic as npm package instead of script
