+++ .env.example
@@ -4,3 +4,8 @@ REACT_APP_TMDB_API_ACCESS_TOKEN =
 REACT_APP_SUPABASE_API_URL =
 REACT_APP_SUPABASE_API_ACCESS_TOKEN =
 REACT_APP_SUPABASE_API_SECRET_TOKEN =
+REACT_APP_NEWRELIC_ACCOUNT_ID =
+REACT_APP_NEWRELIC_AGENT_ID =
+REACT_APP_NEWRELIC_APPLICATION_ID =
+REACT_APP_NEWRELIC_LICENSE_KEY =
+REACT_APP_NEWRELIC_TRUST_KEY =
