+++ package.json
@@ -5,13 +5,12 @@
   "dependencies": {
     "@hookform/resolvers": "^3.1.0",
     "@leandrowkz/tmdb": "^2.1.2",
+    "@newrelic/browser-agent": "^1.237.1",
     "@supabase/supabase-js": "^2.22.0",
-    "@types/md5": "^2.3.2",
     "classnames": "^2.3.2",
     "md5": "^2.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-helmet": "^6.1.0",
     "react-hook-form": "^7.44.3",
     "react-icons": "^4.10.1",
     "react-router-dom": "^6.8.1",
@@ -27,10 +26,10 @@
     "@testing-library/user-event": "^14.0.0",
     "@types/jest": "^29.4.0",
     "@types/jsonp": "^0.2.1",
+    "@types/md5": "^2.3.2",
     "@types/node": "^18.14.0",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
-    "@types/react-helmet": "^6.1.6",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
     "commitlint": "^17.6.1",
     "eslint": "^8.0.1",
